pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contract-template-manager"
version = "0.1.0"
description = "Contract Template Version Control & Approval System"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Noam Raz" },
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-docx>=1.1.0",
    "gitpython>=3.1.41",
    "python-multipart>=0.0.6",
    "streamlit>=1.30.0",
    "requests>=2.31.0",
    "filelock>=3.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.26.0",
    "ruff>=0.2.0",
]

[tool.setuptools.packages.find]
include = ["app*", "frontend*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
